+++ src/SliceZone/index.ts
@@ -1,0 +1,16 @@
+export { TODOSliceComponent } from "./TODOSliceComponent"
+
+export type { DefineComponentSliceComponentProps } from "./getSliceComponentProps"
+export { getSliceComponentProps } from "./getSliceComponentProps"
+
+export { defineSliceZoneComponents } from "./defineSliceZoneComponents"
+
+export type {
+	SliceComponentProps,
+	SliceComponentType,
+	SliceLike,
+	SliceLikeGraphQL,
+	SliceLikeRestV2,
+	SliceZoneComponents,
+	SliceZoneLike,
+} from "./types"
