+++ src/lib/simplyResolveComponent.ts
@@ -5,7 +5,7 @@ import { resolveDynamicComponent } from "vue"
  * A stricter version of {@link resolveDynamicComponent} that resolves only type
  * {@link VNode} for existing components or provided `string`.
  *
- * @param component - An HTML tag name, a component, or a functional component
+ * @param component - An HTML tag name or a component.
  *
  * @returns Resolved component as a {@link VNode} or provided `string`.
  *
